+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Features:
 
 - Define the `--destroy` option for the `infrastructure` and `environment prepare` commands.
 - Define the `ProjectDependenciesUpdate` and `ProjectDependenciesUpdateAndTest` functions, which define the `dependencies update` command.
+- Implement the `diff` and `filesDiff` method of the `GitService`.
 
 ## v0.9.0 (2023-07-28)